requirements.txt
# itsdangerous
# fastapi-mail
# sqlalchemy
# pandas
# pyotp==2.8.0
# fake-useragent==1.1.3
# python-dotenv==1.0.0
# gspread==5.4.0
# df2gspread==1.0.4
# PyPDF2==3.0.1
# fastapi-cors==0.0.6
# python-decouple==3.8
# passlib==1.7.4
# python-jose==3.3.0
# fastapi_permissions==0.2.7
# pydantic[email]
# gunicorn==21.2.0
# openai==0.28.1
# uvicorn
imageio[pyav]
scikit-image
matplotlib
opencv-python
pillow
aiofiles
fastapi-sessions
python-decouple==3.8
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==37.0.1
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.0
fakeredis==1.7.1
fastapi==0.75.2
fastapi-mail==1.0.8
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.22.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
orjson==3.6.8
packaging==21.3
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.8
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
redis==4.1.4
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.36
starlette==0.17.1
typing_extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
#uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.0
